Make combo fire a HUD/skinnable component
- Move to HUD namespace and implement ISerialisableDrawable
- Delete skin.ini stuff
- Scale the skin editor blueprint by 0.25x by moving the main fire
  effect to an internal drawable that applies a reciprocal scale
- Fix comboAlpha and break time alpha not applying invalidations
  correctly
- Use PositionAndColour vertex shader instead of custom one
- More comments on the alpha fields
