+++ osu.Game/Skinning/SkinConfiguration.cs
@@ -39,7 +39,6 @@ namespace osu.Game.Skinning
             LayeredHitSounds,
             AllowSliderBallTint,
             InputOverlayText,
-            ComboFire,
         }
 
         public static List<Color4> DefaultComboColours { get; } = new List<Color4>
